Python: Reorganize expressions module structure (#4468)
